Basic tilemap done. Wish I could rant to someone about why raylib does what it does with sampler2D. Also Fixed render distance bug.

+++ src/jamTiles.h
@@ -6,7 +6,7 @@
 struct tile {
   u32 type;
   // shader information
-  u32 light;
+  u8 light;
 };
 
 struct world {
